Create pivot commands
