+++ src/main/java/frc/robot/commands/pivot/PivotAimCommand.java
@@ -1,0 +1,34 @@
+package frc.robot.commands.pivot;
+
+import edu.wpi.first.wpilibj2.command.Command;
+
+
+public class PivotAimCommand extends Command {
+
+    public PivotAimCommand() {
+        // each subsystem used by the command must be passed into the
+        // addRequirements() method (which takes a vararg of Subsystem)
+        addRequirements();
+    }
+
+    @Override
+    public void initialize() {
+
+    }
+
+    @Override
+    public void execute() {
+
+    }
+
+    @Override
+    public boolean isFinished() {
+        // TODO: Make this return true when this Command no longer needs to run execute()
+        return false;
+    }
+
+    @Override
+    public void end(boolean interrupted) {
+
+    }
+}
